add migration, add promotions fixtures

+++ src/Entity/Promotion.php
@@ -66,12 +66,12 @@ class Promotion
         return $this;
     }
 
-    public function getPrice(): ?int
+    public function getPrice(): ?float
     {
         return $this->price;
     }
 
-    public function setPrice(int $price): self
+    public function setPrice(float $price): self
     {
         $this->price = $price;
